+++ package.json
@@ -41,6 +41,7 @@
     "bootstrap": "4.3.1",
     "connected-react-router": "6.4.0",
     "lodash": "4.17.15",
+    "multiselect-dropdown-react": "1.0.5",
     "object-assign": "4.1.1",
     "react": "16.8.6",
     "react-bootstrap": "1.0.0-beta.14",
